pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "embiggenator"
version = "0.1.0"
description = "LDAP test data generator for Mattermost testing with rroemhild/docker-test-openldap"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "click>=8.1",
    "faker>=20.0",
    "ldap3>=2.9",
    "pyyaml>=6.0",
]

[project.scripts]
embiggenator = "embiggenator.cli:cli"

[tool.setuptools.packages.find]
include = ["embiggenator*"]

[tool.setuptools.package-data]
embiggenator = ["data/defaults/*.ldif"]

[tool.pytest.ini_options]
testpaths = ["tests"]
